Updated DTOs

+++ models/dto/cart.go
@@ -4,5 +4,5 @@ import "keyboardify-server/models"
 
 type CartDTO struct {
 	Products   []models.ProductInCart
-	TotalPrice uint
+	TotalPrice uint64
 }
